requirements.txt
Django>=5.1,<5.2
Pillow
gunicorn
whitenoise
psycopg[binary]
python-dotenv
djangorestframework
django-cors-headers
djangorestframework-simplejwt
